done
I've made a few components for later use such as EntityAccont component which is actually a normal userAccount component but didn't want to name it use since it can be used for a user and a club so I named it EntityAccount. I've Added a read more text that accepts a title and the actual text that needs read more. Ive also added the table that is commonly use in the details screen, whether univercsities, clubs, or even events screens.