Event details screen (#10)
* init

* Revert "init"

This reverts commit ac13e231997c553fe26f6d9e37954f8b26d7e580.

* Revert "Revert "init""

This reverts commit 299986fa460fa3ee4067281314b6aeac6f52a788.

* done

I've made a few components for later use such as EntityAccont component which is actually a normal userAccount component but didn't want to name it use since it can be used for a user and a club so I named it EntityAccount. I've Added a read more text that accepts a title and the actual text that needs read more. Ive also added the table that is commonly use in the details screen, whether univercsities, clubs, or even events screens.

* Update TextWithReadMore.jsx

* bug fixes